requirements.txt
kafka==1.3.5
mysql-connector-python-rf==2.2.2
PyYAML==3.12
redis==2.10.6
redis-py-cluster==1.3.4
requests==2.18.4
apscheduler=3.4.0
selenium==3.11.0
